refactor: Remove qrCodeUrl from Channel and related code
also relocate QR code building into ShowChannelActivity
